[fix] Fix ranking system and Link main-service and ranking-service

See merge request s11-webmobile1-sub2/S11P12A806!201

+++ backend-ranking/src/main/java/com/apink/rankingservice/util/RankCalculator.java
@@ -20,7 +20,7 @@ public class RankCalculator {
         // 기본 점수 계산
         double viewScore = popup.getViews() * VIEW_WEIGHT;
         double likeScore = popup.getLikes() * LIKE_WEIGHT;
-        double reviewScore = calculateBayesianAverage(popup.getReviewScore(), popup.getReviewCount());
+        double reviewScore = calculateBayesianAverage(popup.getReviewScore(), Math.toIntExact(popup.getReviewCount()));
         double reservationScore = popup.getReservations() * RESERVATION_WEIGHT;
 
         // 총점 계산
